patch: move semver script to lib/hnvm, update imports
