+++ lib/hnvm/find-matching-version.js
@@ -1,5 +1,5 @@
 const fs = require('fs')
-const semver = require('../../third-party/semver.js')
+const semver = require('./semver.js')
 
 function isValidSemverRange(range) {
   try {
